streaming/producer/requirements.txt
kafka-python==2.0.2
faker==24.11.0
